refactor(tests): reorganize unit test structure and namespaces
- Moved test files into a new Features folder to better align with application structure.
- Updated namespaces in test files to reflect new folder hierarchy.
- Renamed test classes and files for consistency and clarity (e.g., GetByIdHandlerTests → GetAddressByIdHandlerTests).
- Moved shared test data to Common/TestData.cs for reuse.
- Updated csproj to include new Features folder.
- Improved overall maintainability and discoverability of unit tests.
