+++ AirlineBookingSystem.UnitTests/AirlineBookingSystem.UnitTests.csproj
@@ -28,4 +28,8 @@
       <ProjectReference Include="..\AirlineBookingSystem.Domain\AirlineBookingSystem.Domain.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Features\" />
+    </ItemGroup>
+
 </Project>
