+++ AirlineBookingSystem.UnitTests/Features/Airports/Queries/GetAirportByCodeHandlerTests.cs
@@ -1,12 +1,11 @@
 using AirlineBookingSystem.Application.Features.Airports.Queries.ByCode;
-using AirlineBookingSystem.Application.Features.Airports.Queries.ById;
 using AirlineBookingSystem.Application.Interfaces.Repositories;
 using AirlineBookingSystem.Domain.Entities;
 using AirlineBookingSystem.Shared.DTOs.Airports;
 using AutoMapper;
 using Moq;
 
-namespace AirlineBookingSystem.UnitTests.Airports;
+namespace AirlineBookingSystem.UnitTests.Features.Airports.Queries;
 
 public class GetAirportByCodeHandlerTests
 {
